Refactor to improve testing speed and clarity of location
Added Provider pattern to mock out the TCP client and listener.

MTCAdapter can no longer have its port changed on the fly. For AdapterLab that means fixing it to the default port 7878.

Bump NuGet version to 0.3.0 as the inability to change ports is a breaking change.

Added a new AdapterLab.utest project to separate testing concerns. Similarly moved the CuttingTool into the AdapterLab project and its tests to the AdapterLab.utests project.

Asset is now an interface.

Modified namespaces to be consistent.
