+++ Sample/Program.cs
@@ -15,11 +15,7 @@
  */
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Windows.Forms;
-using Microsoft.Win32;
-using System.Diagnostics;
 using System.Runtime.InteropServices;
