+++ src/Adapter/Providers/TcpClient/TcpClientProvider.cs
@@ -1,0 +1,14 @@
+using System;
+using System.IO;
+using System.Net.Sockets;
+
+namespace MTConnect.Adapter.Providers.TcpClient
+{
+    public interface TcpClientProvider : IDisposable
+    {
+         Stream GetStream();
+         Socket Client { get; }
+         bool Connected { get; }
+         void Close();
+    }
+}
