+++ src/Adapter/Providers/TcpListener/TcpListenerProvider.cs
@@ -1,0 +1,14 @@
+using MTConnect.Adapter.Providers.TcpClient;
+using System.Net;
+using System.Net.Sockets;
+
+namespace MTConnect.Adapter.Providers.TcpListener
+{
+    public interface TcpListenerProvider
+    {
+         EndPoint LocalEndpoint { get; }
+         TcpClientProvider AcceptTcpClient();
+         void Start();
+         void Stop();
+    }
+}
